Move some commands from `common` to the handler (#3766)
